Some changes in new branch
I just put some spaces in between the operators so it is easier to read
And I changed the first for loop for the mine generation into a while loop so it makes more sense if you think about it in English.

Also I am not sure what the if statements are for, are they for checking the # of mines surrounding the selection? If so, should that be in another method?